scraper/requirements.txt
playwright==1.40.0
openai==1.3.0
pinecone-client==2.2.4
python-dotenv==1.0.0
beautifulsoup4==4.12.2
asyncio==3.4.3
aiofiles==23.2.0